\docType{package}
\name{surbar-package}
\alias{surbar}
\alias{surbar-package}
\alias{surbarLibrary}
\title{Title surbar package}
\description{
  More about what it does (maybe more than one line)
}
\details{
  Version 1.0: 14 December 2010
}
\author{
  Various folks at MRI
}
\seealso{
  link to other documents
}

